+++ src/EmailAddresses/CHANGELOG.md
@@ -1,2 +1,2 @@
 ﻿1.0.0
-  - EmailAddresses: Initial release with strongly-typed EmailAddress value object
+  - Initial release with strongly-typed EmailAddress value object.